sd-demo/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
diffusers>=0.27.0
transformers>=4.38.0
torch>=2.2.0
accelerate>=0.27.0
Pillow>=10.2.0
